Address issues with e2e tests by having each suite contain only one test file. (#4896)
* Try using Firefox.

* Try running just one e2e test.

* Now try running two tests. Running more than one test in the same suite seems very unstable.

* Group protractor tests into a single file.

* Rearrange protractor tests so there is one test in each suite.

* Break stateEditor out into separate test in order to avoid timeouts.
