+++ .travis.yml
@@ -7,13 +7,7 @@ python:
 - 2.7
 
 addons:
-  firefox: latest
-  apt:
-    sources:
-      - google-chrome
-    packages:
-      - google-chrome-stable
-      - google-chrome-beta
+  chrome: stable
 
 branches:
   only:
@@ -29,13 +23,18 @@ env:
     # - RUN_BACKEND_TESTS=true REPORT_BACKEND_COVERAGE=true
     # TODO(sll): Reinstate this when the load tests are more reliable.
     # - RUN_BACKEND_TESTS=true REPORT_BACKEND_COVERAGE=true EXCLUDE_LOAD_TESTS=false
-    - RUN_E2E_TESTS_MAIN_EDITOR_PROD=true
-    - RUN_E2E_TESTS_EDITOR_FEATURES_PROD=true
-    - RUN_E2E_TESTS_EXTENSIONS_PROD=true
-    - RUN_E2E_TESTS_LIBRARY_PROD=true
-    - RUN_E2E_TESTS_MISC_PROD=true
-    - RUN_E2E_TESTS_USERS_PROD=true
-    - RUN_E2E_TESTS_EMBEDDING_PROD=true
+    - RUN_E2E_TESTS_ACCESSIBILITY=true
+    - RUN_E2E_TESTS_COLLECTIONS=true
+    - RUN_E2E_TESTS_EDITOR_AND_PLAYER=true
+    - RUN_E2E_TESTS_STATE_EDITOR=true
+    - RUN_E2E_TESTS_EDITOR_FEATURES=true
+    - RUN_E2E_TESTS_EMBEDDING=true
+    - RUN_E2E_TESTS_EXTENSIONS=true
+    # TODO(sll): Reinstate this once the tests for the learner dashboard are fixed.
+    # - RUN_E2E_TESTS_LEARNER_DASHBOARD=true
+    - RUN_E2E_TESTS_LIBRARY=true
+    - RUN_E2E_TESTS_USERS=true
+
 matrix:
   allow_failures: []
   fast_finish: true
@@ -87,13 +86,17 @@ script:
 - if [ $RUN_BACKEND_TESTS == 'true' ] && [ $REPORT_BACKEND_COVERAGE == 'true' ] && [ $EXCLUDE_LOAD_TESTS == 'false' ]; then bash scripts/run_backend_tests.sh --generate_coverage_report; fi
 - if [ $RUN_BACKEND_TESTS == 'true' ] && [ $REPORT_BACKEND_COVERAGE == 'false' ] && [ $EXCLUDE_LOAD_TESTS == 'false' ]; then bash scripts/run_backend_tests.sh; fi
 # Run e2e tests in production mode.
-- if [ $RUN_E2E_TESTS_MAIN_EDITOR_PROD == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="mainEditor" --prod_env; fi
-- if [ $RUN_E2E_TESTS_EDITOR_FEATURES_PROD == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="editorFeatures" --prod_env; fi
-- if [ $RUN_E2E_TESTS_EXTENSIONS_PROD == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="extensions" --prod_env; fi
-- if [ $RUN_E2E_TESTS_LIBRARY_PROD == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="library" --prod_env; fi
-- if [ $RUN_E2E_TESTS_MISC_PROD == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="misc" --prod_env; fi
-- if [ $RUN_E2E_TESTS_USERS_PROD == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="users" --prod_env; fi
-- if [ $RUN_E2E_TESTS_EMBEDDING_PROD == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="embedding" --prod_env; fi
+- if [ $RUN_E2E_TESTS_ACCESSIBILITY == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="accessibility" --prod_env; fi
+- if [ $RUN_E2E_TESTS_COLLECTIONS == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="collections" --prod_env; fi
+- if [ $RUN_E2E_TESTS_EDITOR_AND_PLAYER == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="editorAndPlayer" --prod_env; fi
+- if [ $RUN_E2E_TESTS_STATE_EDITOR == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="stateEditor" --prod_env; fi
+- if [ $RUN_E2E_TESTS_EDITOR_FEATURES == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="editorFeatures" --prod_env; fi
+- if [ $RUN_E2E_TESTS_EMBEDDING == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="embedding" --prod_env; fi
+- if [ $RUN_E2E_TESTS_EXTENSIONS == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="extensions" --prod_env; fi
+- if [ $RUN_E2E_TESTS_LEARNER_DASHBOARD == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="learnerDashboard" --prod_env; fi
+- if [ $RUN_E2E_TESTS_LIBRARY == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="library" --prod_env; fi
+- if [ $RUN_E2E_TESTS_USERS == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="users" --prod_env; fi
+
 
 after_success:
 - if [ $RUN_BACKEND_TESTS == 'true' ] && [ $REPORT_BACKEND_COVERAGE == 'true' ]; then codecov; fi
