+++ core/tests/protractor/editorAndPlayer.js
@@ -16,17 +16,18 @@
  * @fileoverview End-to-end tests of the full exploration editor.
  */
 
-var general = require('../protractor_utils/general.js');
+var editor = require('../protractor_utils/editor.js');
 var forms = require('../protractor_utils/forms.js');
+var general = require('../protractor_utils/general.js');
 var users = require('../protractor_utils/users.js');
 var workflow = require('../protractor_utils/workflow.js');
-var editor = require('../protractor_utils/editor.js');
+
 var CollectionEditorPage =
   require('../protractor_utils/CollectionEditorPage.js');
-var ExplorationPlayerPage =
-  require('../protractor_utils/ExplorationPlayerPage.js');
 var CreatorDashboardPage =
   require('../protractor_utils/CreatorDashboardPage.js');
+var ExplorationPlayerPage =
+  require('../protractor_utils/ExplorationPlayerPage.js');
 var LibraryPage = require('../protractor_utils/LibraryPage.js');
 
 describe('Full exploration editor', function() {
@@ -555,6 +556,67 @@ describe('Full exploration editor', function() {
       editor.setInteraction('EndExploration');
       users.logout();
     });
+  });
+
+  afterEach(function() {
+    general.checkForConsoleErrors([]);
+  });
+});
+
+
+describe('Rating', function() {
+  var EXPLORATION_RATINGTEST = 'RatingTest';
+  var CATEGORY_BUSINESS = 'Business';
+  var LANGUAGE_ENGLISH = 'English';
+  var MINIMUM_ACCEPTABLE_NUMBER_OF_RATINGS = 1;
+  var libraryPage = null;
+  var explorationPlayerPage = null;
+
+  beforeEach(function() {
+    libraryPage = new LibraryPage.LibraryPage();
+    explorationPlayerPage = new ExplorationPlayerPage.ExplorationPlayerPage();
+  });
+
+  var addRating = function(userEmail, userName, explorationName, ratingValue) {
+    users.createUser(userEmail, userName);
+    users.login(userEmail);
+    libraryPage.get();
+    libraryPage.playExploration(EXPLORATION_RATINGTEST);
+    explorationPlayerPage.expectExplorationNameToBe(explorationName);
+    explorationPlayerPage.rateExploration(ratingValue);
+    users.logout();
+  };
+
+  it('should display ratings on exploration when minimum ratings have been ' +
+     'submitted', function() {
+    users.createUser('user1@explorationRating.com', 'user1Rating');
+    // Create an test exploration
+    users.login('user1@explorationRating.com');
+    workflow.createAndPublishExploration(
+      EXPLORATION_RATINGTEST, CATEGORY_BUSINESS,
+      'this is an objective', LANGUAGE_ENGLISH);
+    users.logout();
+
+    // Create test users, play exploration and review them after completion
+    for (var i = 0; i < MINIMUM_ACCEPTABLE_NUMBER_OF_RATINGS - 1; i++) {
+      var userEmail = 'NoDisplay' + i + '@explorationRating.com';
+      var username = 'NoDisplay' + i;
+      addRating(userEmail, username, EXPLORATION_RATINGTEST, 4);
+    }
+
+    libraryPage.get();
+    libraryPage.expectExplorationRatingToEqual(EXPLORATION_RATINGTEST, 'N/A');
+
+    var userEmail = 'Display@explorationRating.com';
+    var username = 'Display';
+    addRating(userEmail, username, EXPLORATION_RATINGTEST, 4);
+
+    libraryPage.get();
+    libraryPage.expectExplorationRatingToEqual(EXPLORATION_RATINGTEST, '4.0');
+
+    libraryPage.playExploration(EXPLORATION_RATINGTEST);
+    explorationPlayerPage.expectExplorationRatingOnInformationCardToEqual(
+      '4.0');
   });
 
   afterEach(function() {
