+++ core/tests/protractor/stateEditor.js
@@ -13,16 +13,15 @@
 // limitations under the License.
 
 /**
- * @fileoverview End-to-end tests of the interaction between the player and
- * editor.
+ * @fileoverview End-to-end tests of the state editor.
  */
 
-var general = require('../protractor_utils/general.js');
-var interactions = require('../../../extensions/interactions/protractor.js');
+var editor = require('../protractor_utils/editor.js');
 var forms = require('../protractor_utils/forms.js');
+var general = require('../protractor_utils/general.js');
 var users = require('../protractor_utils/users.js');
 var workflow = require('../protractor_utils/workflow.js');
-var editor = require('../protractor_utils/editor.js');
+
 var ExplorationPlayerPage =
   require('../protractor_utils/ExplorationPlayerPage.js');
 
@@ -222,6 +221,76 @@ describe('State editor', function() {
     users.logout();
   });
 
+  it('should navigate multiple states correctly, with parameters', function() {
+    var explorationPlayerPage = (
+      new ExplorationPlayerPage.ExplorationPlayerPage());
+    users.createUser('user4@parameters.com', 'user4parameters');
+    users.login('user4@parameters.com');
+
+    workflow.createExploration();
+    editor.enableParameters();
+    editor.addExplorationLevelParameterChange('z', 2);
+
+    editor.setStateName('card 1');
+    editor.addParameterChange('a', 2);
+    editor.setContent(forms.toRichText(
+      'Change value of a from {{a}} to'));
+    editor.setInteraction('NumericInput');
+    editor.addResponse(
+      'NumericInput', null, 'card 2', true, 'IsGreaterThan', 0);
+
+    editor.moveToState('card 2');
+    editor.addParameterChange('a', '{{answer}}');
+    editor.addMultipleChoiceParameterChange('b', [3]);
+    editor.setContent(forms.toRichText(
+      'Change value of b from {{b}} to'));
+    editor.setInteraction('NumericInput');
+    editor.addResponse(
+      'NumericInput', null, 'card 3', true, 'IsGreaterThan', 0);
+
+    editor.moveToState('card 3');
+    editor.addParameterChange('b', '{{answer}}');
+    editor.setContent(forms.toRichText(
+      'sum of {{z}} and {{b}} is {{z + b}},' +
+      ' sum of {{a}} and {{b}} is {{a + b}}'));
+    editor.setInteraction(
+      'MultipleChoiceInput',
+      [forms.toRichText('return'), forms.toRichText('complete')]);
+    editor.addResponse('MultipleChoiceInput', null, 'card 2', false,
+      'Equals', 'return');
+    editor.setDefaultOutcome(null, 'final card', true);
+
+    // Setup a terminating state
+    editor.moveToState('final card');
+    editor.setInteraction('EndExploration');
+    editor.saveChanges();
+
+    general.moveToPlayer();
+    explorationPlayerPage.expectContentToMatch(forms.toRichText(
+      'Change value of a from 2 to'));
+    explorationPlayerPage.submitAnswer('NumericInput', 5);
+    explorationPlayerPage.expectContentToMatch(forms.toRichText(
+      'Change value of b from 3 to'));
+    explorationPlayerPage.submitAnswer('NumericInput', 2);
+    explorationPlayerPage.expectContentToMatch(forms.toRichText(
+      'sum of 2 and 2 is 4, sum of 5 and 2 is 7'));
+    explorationPlayerPage.submitAnswer('MultipleChoiceInput', 'return');
+    explorationPlayerPage.expectContentToMatch(forms.toRichText(
+      'Change value of b from 3 to'));
+    explorationPlayerPage.submitAnswer('NumericInput', 5);
+    explorationPlayerPage.expectContentToMatch(forms.toRichText(
+      'sum of 2 and 5 is 7, sum of 5 and 5 is 10'));
+    explorationPlayerPage.submitAnswer('MultipleChoiceInput', 'return');
+    explorationPlayerPage.expectContentToMatch(forms.toRichText(
+      'Change value of b from 3 to'));
+    explorationPlayerPage.submitAnswer('NumericInput', 4);
+    explorationPlayerPage.expectContentToMatch(forms.toRichText(
+      'sum of 2 and 4 is 6, sum of 5 and 4 is 9'));
+    explorationPlayerPage.submitAnswer('MultipleChoiceInput', 'complete');
+    explorationPlayerPage.expectExplorationToBeOver();
+    users.logout();
+  });
+
   it('should add/modify/delete a hint', function() {
     users.createUser('stateEditorUser2@example.com', 'stateEditorUser2');
     users.login('stateEditorUser2@example.com');
@@ -265,6 +334,58 @@ describe('State editor', function() {
     });
 
     editor.saveChanges();
+    users.logout();
+  });
+
+  it('uses hints and solutions in an exploration', function() {
+    var explorationPlayerPage = (
+      new ExplorationPlayerPage.ExplorationPlayerPage());
+    users.createUser('user1@hintsAndSolutions.com', 'hintsAndSolutions');
+
+    // Creator creates and publishes an exploration
+    users.login('user1@hintsAndSolutions.com');
+    workflow.createExploration();
+
+    editor.setStateName('Introduction');
+    editor.setContent(forms.toRichText('What language is Oppia?'));
+    editor.setInteraction('TextInput');
+    editor.addResponse(
+      'TextInput', forms.toRichText('Good job'), 'End', true, 'Equals',
+      'Finnish');
+    editor.setDefaultOutcome(forms.toRichText('Try again'));
+    editor.addHint('Try language of Finland');
+    editor.addSolution('TextInput', {
+      correctAnswer: 'Finnish',
+      explanation: 'Finland language'
+    });
+    editor.moveToState('End');
+
+    editor.setInteraction('EndExploration');
+    editor.saveChanges();
+
+    general.moveToPlayer();
+    explorationPlayerPage.expectContentToMatch(
+      forms.toRichText('What language is Oppia?'));
+    explorationPlayerPage.submitAnswer('TextInput', 'Roman');
+    // We need to wait some time for the hint to activate.
+    general.waitForSystem();
+    general.waitForSystem();
+    general.waitForSystem();
+
+    explorationPlayerPage.viewHint();
+    explorationPlayerPage.clickGotItButton();
+    explorationPlayerPage.submitAnswer('TextInput', 'Greek');
+    // We need to wait some time for the solution to activate.
+    general.waitForSystem();
+    general.waitForSystem();
+    general.waitForSystem();
+
+    explorationPlayerPage.viewSolution();
+    explorationPlayerPage.clickGotItButton();
+    explorationPlayerPage.expectExplorationToNotBeOver();
+    explorationPlayerPage.submitAnswer('TextInput', 'Finnish');
+    explorationPlayerPage.clickThroughToNextCard();
+    explorationPlayerPage.expectExplorationToBeOver();
     users.logout();
   });
